feat: change type of `side` to float
When calling `compute_sdf` we pass the value for the side length
of the square background mesh. Before it was an `int`, however
it can be any (positive) number, therefore `float` is a better
choice.

+++ msh2sdf/msh2sdf.py
@@ -33,13 +33,13 @@ def _create_square_background_mesh(side:int, nseed:int=100) -> ndarray:
     return XB
 
 
-def compute_sdf(X:ndarray, T:ndarray, side:int=5) -> (ndarray, ndarray):
+def compute_sdf(X:ndarray, T:ndarray, side:float=5) -> (ndarray, ndarray):
     """Compute the signed distance function.
 
     Args:
         X (ndarray): Coordinate matrix of the surface.
         T (ndarray): Connectivity matrix of the surface.
-        side (int): Side length of the background mesh.
+        side (float): Side length of the background mesh.
 
     Returns:
         ndarray: Signed distance function over the background mesh.
